+++ lib/constants/app_info.dart
@@ -1,8 +1,8 @@
 /// App version and info constants
 /// Update version here only - all other files will reference this
 class AppInfo {
-  static const String version = '2.0.2';
-  static const String buildNumber = '32';
+  static const String version = '2.0.3';
+  static const String buildNumber = '33';
   static const String fullVersion = '$version+$buildNumber';