+++ lib/models/settings.g.dart
@@ -21,6 +21,10 @@ AppSettings _$AppSettingsFromJson(Map<String, dynamic> json) => AppSettings(
   folderOrganization: json['folderOrganization'] as String? ?? 'none',
   historyViewMode: json['historyViewMode'] as String? ?? 'grid',
   askQualityBeforeDownload: json['askQualityBeforeDownload'] as bool? ?? true,
+  spotifyClientId: json['spotifyClientId'] as String? ?? '',
+  spotifyClientSecret: json['spotifyClientSecret'] as String? ?? '',
+  useCustomSpotifyCredentials:
+      json['useCustomSpotifyCredentials'] as bool? ?? true,
 );
 
 Map<String, dynamic> _$AppSettingsToJson(AppSettings instance) =>
@@ -39,4 +43,7 @@ Map<String, dynamic> _$AppSettingsToJson(AppSettings instance) =>
       'folderOrganization': instance.folderOrganization,
       'historyViewMode': instance.historyViewMode,
       'askQualityBeforeDownload': instance.askQualityBeforeDownload,
+      'spotifyClientId': instance.spotifyClientId,
+      'spotifyClientSecret': instance.spotifyClientSecret,
+      'useCustomSpotifyCredentials': instance.useCustomSpotifyCredentials,
     };